src/app/requirements.txt
dotenv
chainlit
semantic-kernel[all]
openai
azure-cosmos
azure-storage-blob
azure-ai-projects
azure-identity
asyncpg
psycopg2
sqlalchemy
greenlet
opencensus-ext-azure
azure-monitor-opentelemetry